requirements.txt
flask
requests
dnspython
ipwhois
shodan
